Rename packages to match Google conventions
