+++ main-module/src/main/java/com/diffblue/javatest/Additions.java
@@ -1,4 +1,4 @@
-package com.diffblue.java_test;
+package com.diffblue.javatest;
 
 public class Additions {
   /**
