+++ main-module/src/main/java/com/diffblue/javatest/Calc.java
@@ -1,4 +1,4 @@
-package com.diffblue.java_test;
+package com.diffblue.javatest;
 
 public class Calc {
   int div(int a, int b) {
