+++ main-module/src/main/java/com/diffblue/javatest/somestuff/B.java
@@ -1,4 +1,4 @@
-package com.diffblue.java_test.some_stuff;
+package com.diffblue.javatest.somestuff;
 
 public class B {
   public int x;
