+++ main-module/src/test/java/com/diffblue/javatest/SomeStuffTest.java
@@ -1,4 +1,4 @@
-package com.diffblue.java_test;
+package com.diffblue.javatest;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -6,7 +6,7 @@ import org.junit.Test;
 public class SomeStuffTest {
   @Test
   public void hasB0() {
-    com.diffblue.java_test.some_stuff.A a = new com.diffblue.java_test.some_stuff.A();
+    com.diffblue.javatest.somestuff.A a = new com.diffblue.javatest.somestuff.A();
     boolean r = a.hasB();
     Assert.assertEquals(r, false);
   }
